Menu Enhancement (#45)
Improvements to the main menu.

A number of icons are now created: One for each menu item and a "quit" icon.

The position of the icons move around a circle based on the position of the scroll ring.

Had to use the math functions from lvgl due to issues with gcc when running on the nest.

This adds support for icons and this could be expanded upon in the future to allow the addition of more font awesome icons.

This closes #25